refactor(comp/core/pid): migrate to V2 component pattern
Migrates comp/core/pid from V1 (fx.In dependencies struct, Module() in impl) to V2 (Requires/Provides structs, def/impl/fx/mock layout, compdef.Lifecycle).
