+++ comp/core/pid/def/component.go
